Used `subtract` instead of `-=` for `Resources::contains`.
The current logic of `Resources::contains` is using `-=` to remove
the contained resource object and the `-=` resource object will
invoke `validate` which is not necessary as all of the resource
objects are valid.

The fix is to use `subtract` instead of `-=` for
`Resources::contains`, this can make sure there is no validation
and can improve performance of `Resources::contains` for resources
object.

Review: https://reviews.apache.org/r/50553/