Used `add` instead of `+=` for `Resources::filter`.
The current logic of `Resources::filter` is using `+=` to add
the filtered resource object and the `+=` resource object will
invoke `validate` which is not necessary as all of the resource
objects are valid.

The fix is using `add` instead of `+=` for `Resources::filter`,
this can make sure there is no validation and can improve
performance of `Resources::filter` for resources object.

Review: https://reviews.apache.org/r/50557/