Added extra conditions for deciding when to create "NvidiaComponents".
A recent addition to ensure that 'NvidiaVolume::create()' ran as root
broke all non-root tests on GPU machines.  The reason is that we
unconditionally create this volume so long as we detect
'nvml.isAvailable()', which will fail now that we are only allowed to
create this volume if we have root permissions.

We fix this by adding the proper conditions to determine when / if we
should create this volume based on some combination of --containerizer
and --isolation flags.

Review: https://reviews.apache.org/r/50672/