Avoid unnecessary validation during resource subtraction.
When subtract resources finished, we need to call `validate` to
check if the scalar resource is negative so as to remove this
resource if it is negative. This is a bit heavy as we only need
to check if a scalar has become negative.

This patch is clarifying the logic by checking if the resource
is a negative scalar directly.

Review: https://reviews.apache.org/r/50569/