Updated 'NvidiaVolume' to not error out when binary is missing.
Previously, when building the volume, we would error out if a binary
we were trying to add was not found on the host filesystem. However,
these are not the semantics that we want. By design, we list all the
binaries the *may* exist on the filesystem that we want to put in the
volume, not all of the binaries that *must* exist. To this end, we
should simply skip any unfound binaries and move on to the next one
instead of erroring out.

Review: https://reviews.apache.org/r/50762/