Improved performance in addition and subtraction of resources.
Avoid multiple calls to addable and subtractable in the arithmetic
operations in Resources. While adding or subtracting two Resources
objects, check for addable() is done in Resources::add(), and check
for subtractable() is done in Resources::subtract(). Since the
Resource_ class is private and += and -= operators are only called
from within Resources::add() and Resources::subtract(), we do not
need to call addable() or subtractable() again.

This fixes the performance regression introduced by the Resource_ and
now += and -= performance is on a par with the old Resources before
Resource_ was introduced, according to the test
Resources_BENCHMARK_Test.Arithmetic/0.

Review: https://reviews.apache.org/r/50738/