Removed frameworks with suppressed offers from DRFSorter.
This patch removes frameworks with suppressed offers from the sorter to
reduce time spent in sorting. The allocations will remain in the sorter,
so no data is lost and the numbers are still correct. When a framework
revives offers, it will be re-added to the sorter.

Review: https://reviews.apache.org/r/43666/