Updated allocated number of range to 100 for sorter benchmark test.
If setting port range fragments as 16, then the performance of
`AgentAndClientCount/Sorter_BENCHMARK_Test.FullSort/35` does not
improve much after the resource validation was removed when doing
port range `+=` operations.

We should increase the port range fragments to increase the workload
of the benchmark test for sorter port range fragments calculation.

After the port range fragments was increased to 100, we can see that
the performance of this benchmark test was improved by 10x when adding
port resources with and without resource validation.

Review: https://reviews.apache.org/r/50387/