Renamed the filter for tests that depend on "perf".
Previously, any test case that had "Perf" in its name would only be run
when the perf(1) command was installed. This is inconsistent with other
name-based filters for test cases (e.g., "ROOT_" or "CGROUPS_"). It also
meant that the filter matched "Registrar_BENCHMARK_Test.Performance",
which was a bug (the test does not depend on perf).

Instead, use "PERF_" to indicate tests that can only be run when perf(1)
is installed.

Review: https://reviews.apache.org/r/50625/