Fixed the flaky SlaveRecoveryTest.* failures upon TearDown.
During TearDown of ContainerizerTest<slave::MesosContainerizer>,
we attempt to destroy all cgroups that we can find within the
cgroup root. We observed in CI that on some distributions it
appears that the freeze is not instantaneous and the freezer
delay is needed to complete the destroy process. Howevever,
the clock may be paused at this point which can lead to the
TearDown failing to destroy!

There does not appear to be a listener hook to run code before
test tear down, so a one-off Clock::resume here seems to be the
simplest route.

Review: https://reviews.apache.org/r/50915