Updated Linux 'MountInfoTable' entries to be sorted as expected.
Many places in the codebase assume that the mountinfo table is sorted
according to the order: 'parent mount point < child mount point'.

However, in some cases this may not be true if (for example), a parent
mount point (say '/') is remounted to add some extra flags to it.
When this happens, the remounted file system will appear in the
mountinfo table at the point where it was remounted.

We actually encountered this problem in the wild for the case of '/'
being remounted after '/run' was mounted -- causing problems in the
'NvidiaVolume' which assumes the 'parent < child' ordering.

This commit fixes this problem by building the list of MountInfoTable
entries in sorted order when 'read()' is called. An optional flag can
be used to disable sorting produce the the original ordering.

Review: https://reviews.apache.org/r/50763/