Avoid unnecessarily subtracting resources during contains check.
In Resources::contains, we do not need to subtract resources if
the resource is not a persistent volume as other Resource objects
will be merged together into single Resource ojects.

This fix is adding persistent volume check before subtracting
resources, so that normal resources will not be subtracted. This
is a performance improvement.

Review: https://reviews.apache.org/r/50878/