Removed `default` case from `switch` over enum value.
Compilers can typically warn about enum values not handled in a
`switch`. If one adds a `default` case, even if the enum is extended
later, all cases will already be covered.

Instead do not use a `default` case so that we are warned whenever the
definition of the enum and our pattern matching against its value get
out of sync.

Review: https://reviews.apache.org/r/51041/