Fixed memory leak in agent signal handlers.
This leak was introduced on purpose in:
https://reviews.apache.org/r/34016/

When we run tests in repetition, each agent will create a new signal
handler and leak the old one.  This patch frees the old signal handler
and fixes up some related comments, because the behavior of calling the
signal handler helper multiple times is predictable.

Also fixes a related typo.

Review: https://reviews.apache.org/r/51122/