Added extra functions to the 'Launcher' abstraction.
The functions added in this commit are not currently used anywhere in
the code base. However, they are required for some changes we are
making in order to persist container state in cases where an agent
crashes. The changes to use these function will come in subsequent
commits.

Review: https://reviews.apache.org/r/51167/