Allowed clean up unknown containers in cgroups isolator.
It's likely that isolator `cleanup` is called for an unknown container.
For instance, during launch of a container, `prepare` of some other
isolator failed. Another scenario is that the operator changes the
isolation flags, causing `cleanup` to be called for an isolator that
recovers nothing. Therefore, we should ignore the cleanup if the
container is unknown. In any case, the `cleanup` won't be called
multiple times for an isolator.

Review: https://reviews.apache.org/r/51265/