Renamed `JNIMesos` to `V1Mesos` for scheduler shim.
This change renames `JNIMesos`, v1 implementation for the scheduler
shim to `V1Mesos`. `JNIMesos` was non-intuitive for users considering
the implementation was already in the native code for `V0Mesos` too.
Also, it was a bit confusing that `JNIMesos` referred to using the v1
API under the hood.

Review: https://reviews.apache.org/r/51129/