Refactored MesosContainerizer to prepare for nested container support.
This patch simplified the MesosContainerizer logic by introducing
`ContainerConfig` in `Container` struct. This avoids some parameter
passing in the launch path. To prepare for the nested container
support, this patch introduces a private `launch` method which takes a
`ContainerConfig` (instead of ExecutorInfo and TaskInfo) since nested
container might not tie to a task.

Review: https://reviews.apache.org/r/51163/