Removed a NOTE in MesosContainerizer that no longer applies.
Previously, we need to initialize `launchInfos` because we don't yet
have an explicit state for PROVISIONING (MESOS-4878). After we
introduce the PROVISIONING state, it's guaranteed that `launchInfo`
will be set if the container is in PREPARING state. Therefore, the
NOTE and the corresponding code is no longer needed.

Review: https://reviews.apache.org/r/51417