Add offeredResources to Allocator::updateAllocation() API.
Operations are applied to specific offered resources which are a subset
of all resources allocated to the framework.

For RESERVE/UNRESERVE/CREATE/DESTROY it's not necessary to make such
distinction but we need it for future work such as supporting tasks
using shared resources.

Review: https://reviews.apache.org/r/51412/