Add the 'resources' field back to DRFSorter::Total.
This is a partial reversal of 5a63f2a5d2a5fe0e5315a5b8f79b75b99a6a5893.
We used to have this field but it turned out to be unnecessary for the
current use cases so it was removed. We however will need it for the
ongoing work to support tasks using shared resources.

The details can be found in https://reviews.apache.org/r/45961 but in
essence, we now know of a case where the removal of resource quantities
from the sorter depends on the identity. e.g., To determine if the
quantity of a shared resource can be removed from the sorter we need to
know if there are no longer instances of **the same** (hence identity)
shared resource in the sorter.

Review: https://reviews.apache.org/r/51470/