Removed unnecessary sleep in HealthChecker.
After we have refactored the HealthChecker into a library from the
binary, there is no more need to wait before failing the promise to
ensure the message has been sent to the executor. HealthChecker
lifetime is managed by an executor, hence it is their responsibility
not to clean the instance until after the message with the `kill_task`
flag is received.

Review: https://reviews.apache.org/r/52864