Introduced a new protobuf message 'ContainerRecoverInfo'.
This message will be passed to each isolator during 'recover'. We
decided to make this change due to the following reasons:

1. To support nested containers, the parameters to the old
'Isolator::recover' are not sufficient. For example, a top level
checkpointed executor container may contain child containers that are
unknown (i.e., the Launcher hasn't forked yet, but some isolator
already 'prepare' for it). Currently, the checkpointed ContainerState
list and orphans hashset are not enough for isolators to figure out
those unknown nested containers. Therefore, to support nested
containers, we have to change the 'Isolator::recover' interface. We
need an extra parameter 'knownContainers' to 'Isolator::recover'.

2. We decided to create a new protobuf message for 'Isolator::recover'
interface. Due to the above reason, we have to change the interface.
We simply take this opportunity to use a protobuf so that future
changes won't require isolator writers to adjust their isolators.

In this new protobuf message, we keep the existing parameters in the
old 'Isolator::recover' as the first two proto fields so that existing
isolators can be easily updated. The third field represents a full
list of known containers, which will be used by those isolators that
want to support nesting.

Review: https://reviews.apache.org/r/51779/