Cleaned up a comment that was not accurate.
The comment here most likely is referring to the semantics that occur
at the single call site in the code base if this function returns an
error. However, since the function itself is not the one responsible
for doing what the comment says, I've removed the comment.

Review: https://reviews.apache.org/r/51273/