Removed tasks belonging to a task group from queued tasks.
When sending the `LAUNCH_GROUP` event to the executor, we were
not removing the tasks present in the task group from queued
tasks. This resulted in the tasks being erroneously still
being present in queued tasks thereby not able to receive
terminal status updates from the executor.

Review: https://reviews.apache.org/r/51884/