Added a basic default executor implementation.
This change adds a basic default executor implementation that
just sends a TASK_RUNNING update followed by a TASK_FINISHED
update upon the receipt of a LAUNCH_GROUP event from the executor.
This allows us to test the entire workflow for now. We can iterate
upon this to later send LAUNCH_CONTAINER/DESTROY_CONTAINER
to the agent to actually launch the tasks in a sub container.

Review: https://reviews.apache.org/r/51885/