Made the 'disk/du' isolator nesting aware.
The disk space check will be done at the top container level. All
nested container will not have separate disk space check for now.

In theory, this change is not necessary for now as we can simply mark
the isolator as nesting unaware. However, in the future, we might want
to do separate disk check for each nested container and report
limitations for them. Therefore, we include this patch as a starting
point.

Review: https://reviews.apache.org/r/51989