Updated 'cgroups' isolator 'watch' to be properly nested aware.
In a previous iteration of attempting to make this isolator nested
aware, we (incorrectly) decided to return a 'Failure()' in 'watch()'
for nested containers. However, we shouldn't just blindly fail this
call for nested containers. Instead we should just return a 'Future'
that will never be satisfied. This way it is safe to call 'watch()'
for nested containers, without any unwanted side effects.

This commit fixes this issue.

Review: https://reviews.apache.org/r/51988/