Added test for HTTP endpoints during slave removal.
When a slave is being removed but before the registry operation to
remove it has completed, querying the HTTP endpoints at the master
should not indicate that the slave has been removed yet. This is
important because the registry update might fail; if the master
fails over, the new master will not have removed the slave.

Review: https://reviews.apache.org/r/51377/