Tweaked logging for UnregisterSlaveMessage handler.
Remove a redundant log line: rather than logging when we receive
an UnregisterSlaveMessage, it is sufficient to log how we handle
that message. Also ensure that we log when ignoring an agent
unregister message for an unknown agent, and do some minor code
cleanup.

Review: https://reviews.apache.org/r/51706/