Fixed races between "unreachable" and "unregister" slave transitions.
Now that we update the registry before updating the master's state when
performing these transitions, it is possible for the master to already
be marking a slave unreachable when an `UnregisterSlaveMessage` is
received, or vice versa. Detect these situations and ignore whichever
transition is triggered second.

Review: https://reviews.apache.org/r/51707/