Replaced unchecked `dynamic_cast` with `static_cast`.
`Process` is a CRTP class where `T` must be a valid derived class.
Also, it is customary to use `static_cast` for CRTP methods.

This change uses a `static_cast` instead of a `dynamic_cast`.
`dynamic_cast` can result in a `nullptr` which should be checked.
Here this is not required, so `static_cast` is good enough.

Review: https://reviews.apache.org/r/52025/