Modified the `network/cni` isolator to be nesting aware.
The network file setup in the `network/cni` isolator is now nesting
aware. Since the children share the network and UTS namespace with the
parent, the network files need to be created only for the parent
container. For the child containers, the network files will be simply
a bind mount of the parents network files.

Review: https://reviews.apache.org/r/51857/