Improve DRF sorter performance by bypassing `Resources`.
Currently in `DRFSorter::calculateShare()` (which is called very
frequently), the use of `Resources::get<Scalar>()` is expensive
as it needs to loop over the `Resource` objects and do string
comparison on the `Resource::name` strings.

This patch avoids using `Resources::get<Scalar>()` in
`DRFSorter::calaculateShare()` in favor of maintaining
a map of resource names to scalars.

Note that we had to maintain both this new map and the
previously added `strippedScalarQuantities` since the
latter stores resource roles.

Review: https://reviews.apache.org/r/43665/