Removed resources field from LaunchNestedContainer agent::Call.
This field is currently not used, in that nested containers will
share the resources allocated to the parent container, regardless
of whether resources are specified.

In the future when we introduce the ability to provide nested
container resource allocation and isolation, we can re-introduce
the field as necessary.

Review: https://reviews.apache.org/r/52290