Updated nested container agent API to allow custom ContainerIDs.
Previously, we required that ContainerID.value is a UUID. The
intention here was that each container has the same format, and
we could remove the need for ContainerID.parent since all
ContainerIDs would be unique globally.

With this change, we intend to keep ContainerID.parent and allow
users to specify custom IDs (e.g. <executor uuid>.redis.backup).
This makes for more readable IDs, and we maintain globally unique
ContainerIDs since the executor container has a UUID generated by
the agent.

With this change, we now validate that the user supplied IDs are
setting the parent component of the ID correctly, and that the
value does not contain problematic characters.

Review: https://reviews.apache.org/r/52292