Added `ContainerID` field to the executor `SUBSCRIBED` event.
This change adds the field `ContainerID` to the executor `SUBSCRIBED`
event. This allows the default executor/other custom executors to
then launch nested containers. This is needed as the executors need
to let the agent know of the parent container ID when they make the
`LAUNCH_NESTED_CONTAINER` call. This change also implements the
logic on the agent to send the event to the executor.

Review: https://reviews.apache.org/r/52164/