Removed 'directory' in nested container launch interface.
In this patch, the containerizer launch method for nested containers
should be responsible for figuring out the sandbox directory for the
nested container. So we can avoid passing an extra parameter to the
nested launch interface, and there is no need for the agent to figure
out the nested sandbox.

Review: https://reviews.apache.org/r/52334/