Adjusted the order of using docker config file from URI.
This patch adjusted the order of using the docker config file in
docker containerizer. Now the logic is that we always look for the
docker config file in the sandbox (which may be downloaded by
fetcher via a specified URI). Then, if there does not exist a
docker config file and a docker config file is specified via the
agent flag '--docker_config'. We use that as our new HOME directory.

This should not break previous users since it is most unlikely
any user would specifed both URI and the agent flag for a docker
config file at the same time.

Review: https://reviews.apache.org/r/52372/