Added `pre_exec_command` for top level containers to bring up loopback.
For new network namespaces, pods expect the loop back interface to be
up and the `localhost` to point to this loopback. The loopback
interface is created when a new network namespace is created it is not
brought up. Hence, when a new network namespace is created we
associate a `pre_exec_command` with the container creating this new
network namespace to bring up the loopback. Currently the
`pre_exec_command` is set to "ifconfig lo up".

Review: https://reviews.apache.org/r/52375/

*** Modified for 1.0.x ***