Created symlink from child container to task on the default executor.
We currently don't expose the mapping from the child `ContainerId`
to the `TaskID` of the task on the agent. This makes it very
difficult for the WebUI/other tools to construct the path to the
task sandbox. This change modifies the default executor to create
a symlink from 'tasks/taskId' to the 'containers/containerId'
folder thereby allowing these tools to look up the task sandbox
in the 'tasks' folder.

Review: https://reviews.apache.org/r/52376/