Prevented the allocator from potentially double accounting resources.
Currently, `addSlave()` only adds the used resources of a framework
if the framework is currently registered. The same is not true for
`addFramework()`: prior to this patch, allocator added used
resources of a framework even if the agent is not registered yet.

This is currently not broken because Mesos master populates
frameworks' used resources based on the registered agents only.
However, if we were to call `addFramework()` with resources from
unregistered agents (i.e. for which `addSlave()` had not been
called), double accounting of resources would occur.

Review: https://reviews.apache.org/r/43105/