Fixed the MesosContainerizer destroy issue.
When doing recursive destroy, we should return the collected future of
nested container destroys. Intead, we should fail the corresponding
termination and return that termination if nested container destroys
failed.

In addition, we cannot remove 'Container' struct from the internal map
when the destroy of a nested container failed. This is to ensure that
the top level container do not proceed with destroy if any of its
nested container failed to destroy.

Review: https://reviews.apache.org/r/52444