Ignored error in posix launcher destroy if the container is unknown.
This is consistent to what we do in linux launcher. A container can be
destroyed in launcher but not yet fully cleaned up in isolators. If
agent restarts, we should still allow the cleanup of those containers.
So launcher should ignore destroy request for unknown containers.

Review: https://reviews.apache.org/r/52459