Updated 'destroy()' to checkpoint termination state of nested container.
Previously, when a nested container was being destroyed, it's runtime
directory was being deleted (just the same as a top-level container).
However, this meant that calling 'wait()' on a previously terminated
nested container would return 'None()' since its status had already
been reaped. The problem with this, however, is that this will cause
an entire pod to be terminated since it thinks that the container it
is calling wait on cannot be found.

To fix this, we leave the runtime directory of nested containers
around until their top-level containers are destroyed. Additionally,
we checkpiont the entire termination state of the nested container
into its runtime directory, so that subsequent calls to 'wait()' can
retrieve the full termination state for the lifetime of the top-level
container.