Fixed memory leak in master during framework teardown.
The `process::http::Pipe` class leaks its underlying `shared_ptr`
due to how the master (accidentally) causes the `shared_ptr` to hold
a self-reference.

When the master receives a `SUBSCRIBE` call from an V1 HTTP API
framework, the master creates a `process::http::Pipe` to manage the
reading and writing in separate locations in the code.  For
synchronization, the read/write ends of the HTTP connection share
some state (via `shared_ptr`).

The master introduces a self-reference via this line in
`Master::addFramework`:
```
  http.closed()
    .onAny(defer(self(), &Self::exited, framework->id(), http));
```

`http.closed()` returns a future managed by the read-end of the `Pipe`.
`http` holds a copy of the write-end of the `Pipe`, which in turn has
a copy of the `shared_ptr`.  Because `http` is passed into the future's
continuation, a copy of `http` will live inside the read-end's future
until the either (a) the continuation is executed or (b) the future
is destructed.

Due to how we manage the future, neither (a) nor (b) are performed:
(a) When the read-end of the `Pipe` is closed, we only complete the
    future if the write-end of the `Pipe` is still open.  During
    framework teardown, the write-end is closed first.
(b) The future in question lives inside a `Promise`, which lives
    inside the `shared_ptr`.  Because the self-reference exists, the
    `shared_ptr` is never destructed; which means the `Promise` and
    future are never destructed either.

---

This patch fixes the leak by making sure the continuation is always
executed (a) or discarded.  It does so by discarding the related
future when the write-end of the `Pipe` is already closed.

Review: https://reviews.apache.org/r/51299/