Improved handling of health check failures within the grace period.
The health check library would ignore all failures until after the end
of the grace period.

This behaviour was misleading. With the changes in this commit, once a
health check succeeds for the first time, the grace period rule for
failures is not be applied any more.

Review: https://reviews.apache.org/r/52357/