Added a TODO for cleaning up the `Mesos` library lifetime issues.
The `Mesos` library may invoke the provided callbacks after `Mesos`
is destructed. This has necessitated callers to bind a shared pointer
to the executor when creating the callbacks. However, this should not
be necessary, and this adds a TODO to describe how to eliminate this
issue.

Review: https://reviews.apache.org/r/52567