Updated the way we bring up loopback for a container on a CNI network.
Earlier we were setting up the `pre_exec_command` for the executor
with `ifconfig lo up`. This had a drawback that it assumes that PATH
set for the executor would have `ifconfig`. Unfortunately, it turns
out in most cases the executors PATH is a subset of the agents path
and hence there is a high probability that the executors shell will
not be able to find `ifconfig`.

Instead we bring up the loopback during the setup of network files in
the `NetworkCniIsolatorSetup` subcommand. The subcommand is run with
the agents PATH variable so chances that ifconfig is not present in
the path are much less.

Review: https://reviews.apache.org/r/52622/