Updated `TaskUpdated` event to send `TaskStatus`.
This change modifies the `TaskUpdated` event in the v1 Operator
API to also send the status of task corresponding to the last
status update acknowledged by the scheduler. This is useful
for some subscribers as they might be interested in additional
fields in `TaskStatus` e.g., `ContainerStatus` etc.

Review: https://reviews.apache.org/r/52347/