Fixed agent to correctly handle kill task of unregistered executor.
This change brings in the same logic we had introduced for driver
based executors to shut them down if all of their tasks have been
killed before the executor could initially register (MESOS-5380).
Note that the executors still need to commit suicide if they notice
a disconnection with the agent before they could subscribe with
the agent or if they did not receive a launch event within a
timeout. See MESOS-5385 for more details.

Review: https://reviews.apache.org/r/52651/