Removed 'recover()' interface in 'ContainerLogger'.
This change arises from the nested container support in Mesos.

Currently, the container logger interface only contains `recover()` and
`prepare()` methods.  The `prepare` is called in the containerizers
prior to launching a container, while `recover` is called during
containerizer recovery.  Both methods take two parameters:
an ExecutorInfo and sandbox directory.

Currently, all known implementations of the `ContainerLogger` are
stateless.  This means the `recover()` function is a noop in all
implementations.  Additionally, recovery needs to be paired with an
associated "cleanup" function to be properly stateful.  To avoid
adding tech debt in the containerizers, this removes the `recover()`
function for now.  We can add it back together with a "cleanup"
function if needed in future.

Review: https://reviews.apache.org/r/52762/