Fixed race in getting metrics for dominant share.
When calculate share for dominant share metrics, the client
may have been removed if the dispatch occurs after the client
is removed but before the metric is removed.

The fix is checking if client exist or not before calculate
share for dominant share metrics.

Review: https://reviews.apache.org/r/52678/