Recover resources when offer is rescinded on DESTROY of shared volume.
When a framework issues a DESTROY of a shared volume, and that volume
is not in use by a running or a pending task, we rescind the pending
offers in which the shared volume is present so that the deleted volume
is not assigned to any task in a future ACCEPT call.

At that time, we need to recover the resources as well for proper
accounting of such resources by the allocator.

Review: https://reviews.apache.org/r/52288/